backend/requirements.txt
flask
pyjwt
datetime
uuid
Flask-SQLAlchemy
python-dotenv
werkzeug
